cleaning up the model modules
